# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is a sample server test project.",
      "version": "1.0.0",
      "title": "Swagger BeyondSeek",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "email": "apiteam@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "bs.swagger.io",
    "basePath": "api/v1",
    "schemes": [
      "https"
    ],
    "paths": {
      "/users": {
        "post": {
          "summary": "Add a user to the system",
          "description": "",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "To create a user with username",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid input"
            },
            "201": {
                "description": "Created"
              }
          }
        }
      },
      "/users/{id}/cryptoassociations": {
        "post": {
          "summary": "Create a association onto the system",
          "description": "",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of user",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "name": "cryptoName",
              "in": "body",
              "description": "Updated name of the pet",
              "required": true,
              "type": "string"
            },
            {
              "name": "walletAddress",
              "in": "body",
              "description": "Updated status of the pet",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "201": {
              "description": "Success"
            },
            "400": {
              "description": "Error Returned"
            }
          },
          "security": [
            {
                "bs_auth": []
            }
          ]
        }
      },
      "/users/{id}/wallets": {
        "get": {
          "summary": "Gets the wallet associated with a user",
          "description": "",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/UserWallet"
              }
            }
          },
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of user",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "security": [
              {
                  "bs_auth": []
              }
          ]
        }
      },
      "/login": {
        "post": {
          "summary": "Logs user into the system",
          "description": "",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "username",
              "in": "query",
              "description": "The user name for login",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "query",
              "description": "The password for login in clear text",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "string"
              }
            },
            "400": {
              "description": "Invalid username/password supplied"
            }
          }
        }
      }
    },
    "securityDefinitions": {
      "bs_auth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UserWallet" : {
        "type": "array",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "UserWallets": {
            "type" : "array",
            "properties" : {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "cryptoName": {
                "type": "string"
              },
              "userid": {
                "type": "integer",
                "format": "int64"
              },
              "walletAddress": {
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    }
  }